Fix F22 Cockpit view too low
Jorge (aka japreja):
I was struggling with the cockpit view and figure out how to fix it.
Just so you know the “Adjust View Position” controlls are way too 
sensitive in version 3.7 (and previous versions).
 
here is the diff from tortoise SVN for the F-22 from 
fgaddon “Aircraft/Lockheed-Martin-FA-22A-Raptor”


git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Lockheed-Martin-FA-22A-Raptor@819 9eb9bc84-db9d-4e05-893b-26dd41ac94a6